TracMasterTickets 4.0.4dev: Allow image format to be specified
The first entry in `[mastertickets]` `acceptable_formats`
will be used by default. The PNG format will be used if the
list is empty. The new default value for `acceptable_formats`
is `svg,png,cmapx`, so SVG will be the default format
if `acceptable_formats` is not specified in the
`[mastertickets]` section of trac.ini.

Fixes #13631.


git-svn-id: https://trac-hacks.org/svn/masterticketsplugin@17716 7322e99d-02ea-0310-aa39-e9a107903beb